+++ app/assets/javascripts/3dbio_viewer/src/domain/utils/i18n.ts
@@ -1,3 +1,9 @@
-import i18n from "../../webapp/utils/i18n";
+import i18n from "d2-ui-components/locales";
 
-export default i18n;
+const i18nWrapper = {
+    t(s: string, namespace?: object): string {
+        return i18n.t(s, namespace);
+    },
+};
+
+export default i18nWrapper;
